Moved collect storage usage to separate file and class
